# Release v0.3.4.2b 2025-10-27
## New Features

1. Added server_port configuration, 0 disables server if required (v0.3.4.1b 2025-10-01)
2. Improve client replication io (v0.3.4.1b 2025-10-07)
3. Make server replication asynch and faster (v0.3.4.1b 2025-10-08)
4. server replication can be called multiple times per connection (v0.3.4.1b 2025-10-08)
5. add shard info `INFO SHARD [key]|[#(shard number)]` (v0.3.4.2b 2025-10-17)
   - `INFO SHARD 1` returns shard info for key `1`
   - `INFO SHARD #1` returns shard info for shard number 1
   - `INFO SHARD ##1` or `#ABC` returns shard info for key `##1` or `#ABC`
6. add key space support with these new api's key space name's should match `[0-9,A-Z,a-z,_]+`
    - `USE ksname` the named keyspace is made the current store for all other API's
    - `UNLOAD ksname` all memory used by named key space is cleared except if it's in use
    - `SPACES` returns a list of loaded key spaces and their sizes
    - total count in the entire db of all shards in all key spaces
      `SIZEALL`
    - save and clear all key spaces
      `CLEARALL`
      `SAVEALL`
    - `KSOPTION SET [UNORDERED|ORDERED]` sets the current key space to ordered or unordered, option is saved in key space shards
7. add optional extra part to storage for future backwards compatibility

+++ src/hash_arena.cpp
@@ -64,7 +64,7 @@ bool arena::base_hash_arena::save(const std::string &filename,
 
 
 bool arena::base_hash_arena::send(std::ostream &out, const std::function<void(std::ostream &)> &extra, bool write_version) const {
-    uint64_t completed = write_version ? storage_version : 0;
+    uint64_t completed = write_version ? (int)storage_version : 0;
     size_t size = hidden_arena.size();
     writep(out, completed);
     writep(out, max_address_accessed);
